refactor: extract history into a standalone file and fix eslint config
