+++ eslint.config.mjs
@@ -7,12 +7,22 @@ export default defineConfig([
   { files: ["**/*.{js,mjs,cjs}"] },
   {
     files: ["**/*.{js,mjs,cjs}"],
-    languageOptions: { globals: globals.browser, ...globals.webextensions },
+    languageOptions: {
+      globals: {
+        ...globals.browser,
+        ...globals.webextensions,
+        jQuery: 'readonly',
+        $: 'readonly',
+      },
+    },
   },
   {
     files: ["**/*.{js,mjs,cjs}"],
     plugins: { js },
     extends: ["js/recommended"],
+    rules: {
+      'no-unused-vars': 'warn', // Change 'error' to 'warn' for unused variables
+    },
   },
   eslintConfigPrettier,
 ]);
